Keep the cookie implementation behind the login flow.
Addresses review: group the login-flow modules, and put a lid on the
cookie mechanics.

The RFC 6265 handling is not specific to this flow, but nothing else
uses it, so it now lives inside cookieCapture with no export — behind
the LoginFlow interface, where nobody needs to care how this one flow
works. A note there says to lift it back out if cookie handling is ever
wanted more generally, which is how it got its own module in the first
place.

Its tests move up with it, restated against the flow: a cookie's
default path, Domain matching with and without the leading dot, the
three ways a cookie is cleared, replacement before the login finishes,
and a future expiry are all observable in the header the flow ends up
storing.

The flow's parameter schema stopped being exported for the same reason.
What cookieCapture.ts offers the rest of the codebase is one class.

The three login-flow modules move to services/core/loginFlows/ as
base, registry and cookieCapture.
