+++ src/services/core/registered.ts
@@ -9,7 +9,7 @@
 
 import { ApiCredentialStatus, type ApiCredentials } from '../../apiCredentials/base.js';
 import { Service, type ServiceSession } from './base.js';
-import type { LoginFlow } from './loginFlow.js';
+import type { LoginFlow } from './loginFlows/base.js';
 
 /**
  * Where a registered service gets its browser login, if anywhere. The two
